added applicayion ymlConfig file, moved discordBotBootstrap to infra
