+++ src/main/java/dev/salonce/discordquizbot/infrastructure/configs/ApplicationYmlConfig.java
@@ -1,0 +1,14 @@
+package dev.salonce.discordquizbot.infrastructure.configs;
+
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
+
+public final class ApplicationYmlConfig {
+    public static void configure(){
+        Resource primaryResource = new ClassPathResource("private/application.yml");
+        String configLocation = primaryResource.exists() ?
+                "classpath:/private/application.yml" :
+                "classpath:/sample/application-sample.yml";
+        System.setProperty("spring.config.location", configLocation);
+    }
+}
